Backend/app/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0

# AI & ML
openai==1.12.0
langchain-openai==0.3.32
langchain-core==0.3.75
langgraph>=0.2.0

# Document Processing & RAG
sentence-transformers==2.2.2  # Convert text to vectors
pypdf2==3.0.1  # Read PDF files
python-docx==1.1.0  # Read Word files
numpy>=1.21.0  # For vector calculations

# Data & Utilities
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging
structlog==23.2.0

# Tavily API
tavily-python==0.3.0
langchain-tavily>=0.1.0